Fix backpropagation and forward pass
The network can now overfit to a simple linear regression problem, which means that is sort of does the right thing.

There's still this weird error when it sometimes starts going to the wrong direction, and ends up printing nan and inf. But in most cases it can actually learn stuff.

Also added tests using pytorch.